Repair run target
The target depended on build which was removed in
b4d471ed4a8b28a048fcf62eb79eeffe0d1e9ba8 to compile the binary.
Now the run target requires on quick which builds the binary as well.
Also, I removed the redundant cding into cmd/rwtxt because the path can
be passed to go build.

Note that quick should be renamed to rwtxt because this is the file the
target produces.  But, this change is outside the scope of this PR.